+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 6.4.0 (March 21, 2025)
+
+FEATURES:
+
+* Get endpoint for DNSv2 client from keystone catalog ([#333](https://github.com/selectel/terraform-provider-selectel/pull/333))
+
+## 6.3.0 (March 12, 2025)
+
+BUG FIXES:
+* Added filters to `quotas.GetProjectQuotas` calls to avoid errors ([#330](https://github.com/selectel/terraform-provider-selectel/pull/330))
+
 ## 6.2.0 (Feb 24, 2025)
 
 FEATURES: